etl-server/requirements.txt
psycopg2-binary
dotenv
pyzmq
wheel
pandas
matplotlib
soda-core-postgres
